+++ src/pczt/prover.rs
@@ -81,14 +81,8 @@ impl super::Bundle {
                     .clone()
                     .ok_or(ProverError::MissingValueCommitTrapdoor)?;
 
-                Circuit::from_action_context_for_version(
-                    spend,
-                    output_note,
-                    alpha,
-                    rcv,
-                    pk.circuit_version(),
-                )
-                .ok_or(ProverError::RhoMismatch)
+                Circuit::from_action_context(spend, output_note, alpha, rcv, pk.circuit_version())
+                    .ok_or(ProverError::RhoMismatch)
             })
             .collect::<Result<Vec<_>, ProverError>>()?;